app-shack/rootfs/app/requirements.txt
# HA Shim core dependencies
paho-mqtt>=1.6.0
pyyaml>=6.0
colorlog>=6.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
aiohttp>=3.9.0
voluptuous>=0.14.0
python-multipart>=0.0.6
